[CCS-38] Move temp NOTAM classes into final package structure
